+++ pytest_flask/__init__.py
@@ -1,4 +1,2 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-__version__ = "0.10.0"
+from ._version import version as __version__