+++ tests/conftest.py
@@ -1,26 +1,26 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
-import pytest
-
 from textwrap import dedent
-from flask import Flask, jsonify
+
+import pytest
+from flask import Flask
+from flask import jsonify
 
 
-pytest_plugins = 'pytester'
+pytest_plugins = "pytester"
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def app():
     app = Flask(__name__)
-    app.config['SECRET_KEY'] = '42'
+    app.config["SECRET_KEY"] = "42"
 
-    @app.route('/')
+    @app.route("/")
     def index():
-        return app.response_class('OK')
+        return app.response_class("OK")
 
-    @app.route('/ping')
+    @app.route("/ping")
     def ping():
-        return jsonify(ping='pong')
+        return jsonify(ping="pong")
 
     return app
 
@@ -28,23 +28,26 @@ def app():
 @pytest.fixture
 def appdir(testdir):
     app_root = testdir.tmpdir
-    test_root = app_root.mkdir('tests')
+    test_root = app_root.mkdir("tests")
 
-    def create_test_module(code, filename='test_app.py'):
+    def create_test_module(code, filename="test_app.py"):
         f = test_root.join(filename)
         f.write(dedent(code), ensure=True)
         return f
 
     testdir.create_test_module = create_test_module
 
-    testdir.create_test_module('''
+    testdir.create_test_module(
+        """
         import pytest
 
         from flask import Flask
 
-        @pytest.fixture
+        @pytest.fixture(scope='session')
         def app():
             app = Flask(__name__)
             return app
-    ''', filename='conftest.py')
+    """,
+        filename="conftest.py",
+    )
     return testdir